module3-api-service/requirements.txt
# Django Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
drf-spectacular==0.26.5

# Database
psycopg2-binary==2.9.9
django-extensions==3.2.3

# Geospatial
GDAL==3.7.3
Shapely==2.0.2

# Authentication
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0

# WebSocket Support
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1

# HTTP Client
aiohttp==3.9.1
requests==2.31.0

# Background Tasks
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Monitoring
prometheus-client==0.19.0
sentry-sdk==1.38.0

# Development
pytest==7.4.3
pytest-django==4.7.0
pytest-asyncio==0.21.1
factory-boy==3.3.0
coverage==7.3.2

# Production
gunicorn==21.2.0
whitenoise==6.6.0
